% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.securityhub_operations.R
\name{accept_invitation}
\alias{accept_invitation}
\title{Accepts the invitation to be monitored by a master SecurityHub account}
\usage{
accept_invitation(
  MasterId = "string",
  InvitationId = "string"
)
}
\arguments{
\item{MasterId}{The account ID of the master Security Hub account whose invitation you're accepting.}

\item{InvitationId}{The ID of the invitation that is sent to the AWS account by the Security Hub master account.}
}
\description{
Accepts the invitation to be monitored by a master SecurityHub account.
}
\examples{

}
