% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.securityhub_operations.R
\name{batch_disable_standards}
\alias{batch_disable_standards}
\title{Disables the standards specified by the standards subscription ARNs}
\usage{
batch_disable_standards(
  StandardsSubscriptionArns = list(
    "string"
  )
)
}
\arguments{
\item{StandardsSubscriptionArns}{[required] The ARNS of the standards subscriptions that you want to disable.}
}
\description{
Disables the standards specified by the standards subscription ARNs. In the context of Security Hub, supported standards (for example, CIS AWS Foundations) are automated and continuous checks that help determine your compliance status against security industry (including AWS) best practices.
}
\examples{

}
