% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.securityhub_operations.R
\name{create_members}
\alias{create_members}
\title{Creates member Security Hub accounts in the current AWS account (which becomes the master Security Hub account) that has Security Hub enabled}
\usage{
create_members(
  AccountDetails = list(
    list(
      AccountId = "string",
      Email = "string"
    )
  )
)
}
\arguments{
\item{AccountDetails}{A list of account ID and email address pairs of the accounts that you want to associate with the master Security Hub account.}
}
\description{
Creates member Security Hub accounts in the current AWS account (which becomes the master Security Hub account) that has Security Hub enabled.
}
\examples{

}
