% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.securityhub_operations.R
\name{decline_invitations}
\alias{decline_invitations}
\title{Declines invitations that are sent to this AWS account (invitee) by the AWS accounts (inviters) that are specified by the account IDs}
\usage{
decline_invitations(
  AccountIds = list(
    "string"
  )
)
}
\arguments{
\item{AccountIds}{A list of account IDs specifying accounts whose invitations to Security Hub you want to decline.}
}
\description{
Declines invitations that are sent to this AWS account (invitee) by the AWS accounts (inviters) that are specified by the account IDs.
}
\examples{

}
