% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.securityhub_operations.R
\name{delete_insight}
\alias{delete_insight}
\title{Deletes an insight that is specified by the insight ARN}
\usage{
delete_insight(InsightArn)
}
\arguments{
\item{InsightArn}{[required] The ARN of the insight that you want to delete.}
}
\description{
Deletes an insight that is specified by the insight ARN.
}
\section{Accepted Parameters}{
\preformatted{delete_insight(
  InsightArn = "string"
)
}
}

