% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.securityhub_operations.R
\name{delete_invitations}
\alias{delete_invitations}
\title{Deletes invitations that are sent to this AWS account (invitee) by the AWS accounts (inviters) that are specified by their account IDs}
\usage{
delete_invitations(AccountIds = NULL)
}
\arguments{
\item{AccountIds}{A list of account IDs specifying accounts whose invitations to Security Hub you want to delete.}
}
\description{
Deletes invitations that are sent to this AWS account (invitee) by the AWS accounts (inviters) that are specified by their account IDs.
}
\section{Accepted Parameters}{
\preformatted{delete_invitations(
  AccountIds = list(
    "string"
  )
)
}
}

