% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.securityhub_operations.R
\name{delete_members}
\alias{delete_members}
\title{Deletes the Security Hub member accounts that are specified by the account IDs}
\usage{
delete_members(AccountIds = NULL)
}
\arguments{
\item{AccountIds}{A list of account IDs of the Security Hub member accounts that you want to delete.}
}
\description{
Deletes the Security Hub member accounts that are specified by the account IDs.
}
\section{Accepted Parameters}{
\preformatted{delete_members(
  AccountIds = list(
    "string"
  )
)
}
}

