% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.securityhub_operations.R
\name{disable_import_findings_for_product}
\alias{disable_import_findings_for_product}
\title{Stops you from being able to import findings generated by integrated third-party providers into Security Hub}
\usage{
disable_import_findings_for_product(
  ProductSubscriptionArn = "string"
)
}
\arguments{
\item{ProductSubscriptionArn}{[required] The ARN of a resource that represents your subscription to a supported product.}
}
\description{
Stops you from being able to import findings generated by integrated third-party providers into Security Hub.
}
\examples{

}
