% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.securityhub_operations.R
\name{enable_import_findings_for_product}
\alias{enable_import_findings_for_product}
\title{Enables you to import findings generated by integrated third-party providers into Security Hub}
\usage{
enable_import_findings_for_product(
  ProductArn = "string"
)
}
\arguments{
\item{ProductArn}{[required] The ARN of the product that generates findings that you want to import into Security Hub.}
}
\description{
Enables you to import findings generated by integrated third-party providers into Security Hub.
}
\examples{

}
