% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.securityhub_operations.R
\name{enable_security_hub}
\alias{enable_security_hub}
\title{Enables the AWS Security Hub service}
\usage{
enable_security_hub()
}
\description{
Enables the AWS Security Hub service.
}
\examples{

}
