% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.securityhub_operations.R
\name{get_insight_results}
\alias{get_insight_results}
\title{Lists the results of the Security Hub insight specified by the insight ARN}
\usage{
get_insight_results(
  InsightArn = "string"
)
}
\arguments{
\item{InsightArn}{[required] The ARN of the insight whose results you want to see.}
}
\description{
Lists the results of the Security Hub insight specified by the insight ARN.
}
\examples{

}
