% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.securityhub_operations.R
\name{get_invitations_count}
\alias{get_invitations_count}
\title{Returns the count of all Security Hub membership invitations that were sent to the current member account, not including the currently accepted invitation}
\usage{
get_invitations_count()
}
\description{
Returns the count of all Security Hub membership invitations that were sent to the current member account, not including the currently accepted invitation.
}
\examples{

}
