% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.securityhub_operations.R
\name{get_master_account}
\alias{get_master_account}
\title{Provides the details for the Security Hub master account to the current member account}
\usage{
get_master_account()
}
\description{
Provides the details for the Security Hub master account to the current member account.
}
\section{Accepted Parameters}{
\preformatted{get_master_account()
}
}

