% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.securityhub_operations.R
\name{get_members}
\alias{get_members}
\title{Returns the details on the Security Hub member accounts that are specified by the account IDs}
\usage{
get_members(
  AccountIds = list(
    "string"
  )
)
}
\arguments{
\item{AccountIds}{[required] A list of account IDs for the Security Hub member accounts on which you want to return the details.}
}
\description{
Returns the details on the Security Hub member accounts that are specified by the account IDs.
}
\examples{

}
