% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.securityhub_operations.R
\name{list_enabled_products_for_import}
\alias{list_enabled_products_for_import}
\title{Lists all Security Hub-integrated third-party findings providers}
\usage{
list_enabled_products_for_import(NextToken = NULL, MaxResults = NULL)
}
\arguments{
\item{NextToken}{Paginates results. Set the value of this parameter to NULL on your first call to the ListEnabledProductsForImport operation. For subsequent calls to the operation, fill nextToken in the request with the value of NextToken from the previous response to continue listing data.}

\item{MaxResults}{Indicates the maximum number of items that you want in the response.}
}
\description{
Lists all Security Hub-integrated third-party findings providers.
}
\section{Accepted Parameters}{
\preformatted{list_enabled_products_for_import(
  NextToken = "string",
  MaxResults = 123
)
}
}

