% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.securityhub_operations.R
\name{list_invitations}
\alias{list_invitations}
\title{Lists all Security Hub membership invitations that were sent to the current AWS account}
\usage{
list_invitations(MaxResults = NULL, NextToken = NULL)
}
\arguments{
\item{MaxResults}{Indicates the maximum number of items that you want in the response.}

\item{NextToken}{Paginates results. Set the value of this parameter to NULL on your first call to the ListInvitations operation. For subsequent calls to the operation, fill nextToken in the request with the value of NextToken from the previous response to continue listing data.}
}
\description{
Lists all Security Hub membership invitations that were sent to the current AWS account.
}
\section{Accepted Parameters}{
\preformatted{list_invitations(
  MaxResults = 123,
  NextToken = "string"
)
}
}

