% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.securityhub_operations.R
\name{disassociate_members}
\alias{disassociate_members}
\title{Disassociates the Security Hub member accounts that are specified by the account IDs from their master account}
\usage{
disassociate_members(AccountIds = NULL)
}
\arguments{
\item{AccountIds}{The account IDs of the member accounts that you want to disassociate from the master account.}
}
\description{
Disassociates the Security Hub member accounts that are specified by the account IDs from their master account.
}
\section{Accepted Parameters}{
\preformatted{disassociate_members(
  AccountIds = list(
    "string"
  )
)
}
}

